+++ documentation/docs/40-best-practices/20-seo.md
@@ -56,97 +56,3 @@ export async function GET() {
 	);
 }
 ```
-
-### AMP
-
-An unfortunate reality of modern web development is that it is sometimes necessary to create an [Accelerated Mobile Pages (AMP)](https://amp.dev/) version of your site. In SvelteKit this can be done by setting the [`inlineStyleThreshold`](configuration#inlineStyleThreshold) option...
-
-```js
-/// file: vite.config.js
-import { defineConfig } from 'vite';
-import { sveltekit } from '@sveltejs/kit/vite';
-
-export default defineConfig({
-	plugins: [
-		sveltekit({
-			// since <link rel="stylesheet"> isn't
-			// allowed, inline all styles
-			inlineStyleThreshold: Infinity
-		})
-	]
-});
-```
-
-...disabling `csr` in your root `+layout.js`/`+layout.server.js`...
-
-```js
-/// file: src/routes/+layout.server.js
-export const csr = false;
-```
-
-...adding `amp` to your `app.html`
-
-```html
-<html amp>
-...
-```
-
-...and transforming the HTML using `transformPageChunk` along with `transform` imported from `@sveltejs/amp`:
-
-```js
-/// file: src/hooks.server.js
-import * as amp from '@sveltejs/amp';
-
-/** @type {import('@sveltejs/kit').Handle} */
-export async function handle({ event, resolve }) {
-	let buffer = '';
-	return await resolve(event, {
-		transformPageChunk: ({ html, done }) => {
-			buffer += html;
-			if (done) return amp.transform(buffer);
-		}
-	});
-}
-```
-
-To prevent shipping any unused CSS as a result of transforming the page to amp, we can use [`dropcss`](https://www.npmjs.com/package/dropcss):
-
-```js
-// @filename: ambient.d.ts
-declare module 'dropcss';
-
-// @filename: index.js
-// ---cut---
-/// file: src/hooks.server.js
-// @errors: 2307
-import * as amp from '@sveltejs/amp';
-import dropcss from 'dropcss';
-
-/** @type {import('@sveltejs/kit').Handle} */
-export async function handle({ event, resolve }) {
-	let buffer = '';
-
-	return await resolve(event, {
-		transformPageChunk: ({ html, done }) => {
-			buffer += html;
-
-			if (done) {
-				let css = '';
-				const markup = amp
-					.transform(buffer)
-					.replace('⚡', 'amp') // dropcss can't handle this character
-					.replace(/<style amp-custom([^>]*?)>([^]+?)<\/style>/, (match, attributes, contents) => {
-						css = contents;
-						return `<style amp-custom${attributes}></style>`;
-					});
-
-				css = dropcss({ css, html: markup }).css;
-				return markup.replace('</style>', `${css}</style>`);
-			}
-		}
-	});
-}
-
-```
-
-> [!NOTE] It's a good idea to use the `handle` hook to validate the transformed HTML using `amphtml-validator`, but only if you're prerendering pages since it's very slow.
