--- packages/amp/CHANGELOG.md
@@ -1,92 +1,0 @@
-# @sveltejs/amp
-
-## 1.1.5
-### Patch Changes
-
-
-- chore: add `.git` to the end of `package.json` repository url ([#14134](https://github.com/sveltejs/kit/pull/14134))
-
-- Updated dependencies [[`c968aef`](https://github.com/sveltejs/kit/commit/c968aef5727f978244d5160657b4a7ac651384ae)]:
-  - @sveltejs/kit@2.27.3
-
-## 1.1.4
-### Patch Changes
-
-
-- docs: update URLs for new svelte.dev site ([#12857](https://github.com/sveltejs/kit/pull/12857))
-
-- Updated dependencies [[`dcbe4222a194c5f90cfc0fc020cf065f7a4e4c46`](https://github.com/sveltejs/kit/commit/dcbe4222a194c5f90cfc0fc020cf065f7a4e4c46), [`4cdbf76fbbf0c0ce7f574ef69c8daddcf954d39d`](https://github.com/sveltejs/kit/commit/4cdbf76fbbf0c0ce7f574ef69c8daddcf954d39d), [`3a9b78f04786898ca93f6d4b75ab18d26bc45192`](https://github.com/sveltejs/kit/commit/3a9b78f04786898ca93f6d4b75ab18d26bc45192), [`723eb8b31e6a22c82f730c30e485386c8676b746`](https://github.com/sveltejs/kit/commit/723eb8b31e6a22c82f730c30e485386c8676b746), [`8ec471c875345b751344e67580ff1b772ef2735b`](https://github.com/sveltejs/kit/commit/8ec471c875345b751344e67580ff1b772ef2735b)]:
-  - @sveltejs/kit@2.7.3
-
-## 1.1.3
-### Patch Changes
-
-
-- chore: configure provenance in a simpler manner ([#12570](https://github.com/sveltejs/kit/pull/12570))
-
-- Updated dependencies [[`087a43d391fc38b8c008fb39a804dc6988974101`](https://github.com/sveltejs/kit/commit/087a43d391fc38b8c008fb39a804dc6988974101)]:
-  - @sveltejs/kit@2.5.22
-
-## 1.1.2
-### Patch Changes
-
-
-- chore: package provenance ([#12567](https://github.com/sveltejs/kit/pull/12567))
-
-- Updated dependencies [[`4930a8443caa53bcecee7b690cd28e429b1c8a20`](https://github.com/sveltejs/kit/commit/4930a8443caa53bcecee7b690cd28e429b1c8a20)]:
-  - @sveltejs/kit@2.5.21
-
-## 1.1.1
-
-### Patch Changes
-
-- chore: add keywords for discovery in npm search ([#12330](https://github.com/sveltejs/kit/pull/12330))
-
-- Updated dependencies [[`25acb1d9fce998dccd8050b93cf4142c2b082611`](https://github.com/sveltejs/kit/commit/25acb1d9fce998dccd8050b93cf4142c2b082611), [`642c4a4aff4351b786fe6274aa2f0bf7d905faf9`](https://github.com/sveltejs/kit/commit/642c4a4aff4351b786fe6274aa2f0bf7d905faf9), [`0a0e9aa897123ebec50af08e9385b2ca4fc5bb28`](https://github.com/sveltejs/kit/commit/0a0e9aa897123ebec50af08e9385b2ca4fc5bb28)]:
-  - @sveltejs/kit@2.5.11
-
-## 1.1.0
-
-### Minor Changes
-
-- feat: allow SvelteKit 2 as peer dependency ([#11233](https://github.com/sveltejs/kit/pull/11233))
-
-## 1.0.2
-
-### Patch Changes
-
-- fix: publish missing files ([#8532](https://github.com/sveltejs/kit/pull/8532))
-
-## 1.0.1
-
-### Patch Changes
-
-- chore: remove superfluous main field from package.json ([#8519](https://github.com/sveltejs/kit/pull/8519))
-
-- Updated dependencies [[`7e2d3405`](https://github.com/sveltejs/kit/commit/7e2d34056e99f371e22406d941b764df365a2649)]:
-  - @sveltejs/kit@1.1.1
-
-## 1.0.0
-
-### Major Changes
-
-First major release, see below for the history of changes that lead up to this.
-Starting from now all releases follow semver and changes will be listed as Major/Minor/Patch
-
-## 1.0.0-next.3
-
-### Patch Changes
-
-- chore: add peerDependencies, add more specific next version ([#8141](https://github.com/sveltejs/kit/pull/8141))
-
-## 1.0.0-next.2
-
-### Patch Changes
-
-- Amend boilerplate to match current recommendations ([#7184](https://github.com/sveltejs/kit/pull/7184))
-
-## 1.0.0-next.1
-
-### Patch Changes
-
-- Add types to pkg.exports ([#5045](https://github.com/sveltejs/kit/pull/5045))
