--- packages/amp/index.d.ts
@@ -1,1 +1,0 @@
-export function transform(html: string): string;
