--- packages/amp/tsconfig.json
@@ -1,10 +1,0 @@
-{
-	"compilerOptions": {
-		"allowJs": true,
-		"checkJs": true,
-		"noEmit": true,
-		"target": "es2022",
-		"module": "node16",
-		"moduleResolution": "node16"
-	}
-}
