--- packages/kit/test/apps/amp/src/hooks.server.js
@@ -1,27 +1,0 @@
-import * as amp from '@sveltejs/amp';
-import dropcss from 'dropcss';
-
-/** @type {import('@sveltejs/kit').Handle} */
-export async function handle({ event, resolve }) {
-	let buffer = '';
-
-	return await resolve(event, {
-		transformPageChunk: ({ html, done }) => {
-			buffer += html;
-
-			if (done) {
-				let css = '';
-				const markup = amp
-					.transform(buffer)
-					.replace('⚡', 'amp')
-					.replace(/<style amp-custom([^>]*?)>([^]+?)<\/style>/, (match, attributes, contents) => {
-						css = contents;
-						return `<style amp-custom${attributes}></style>`;
-					});
-
-				css = dropcss({ css, html: markup }).css;
-				return markup.replace('</style>', `${css}</style>`);
-			}
-		}
-	});
-}
