--- packages/kit/test/apps/amp/src/routes/http-equiv/cache-control/+page.js
@@ -1,8 +1,0 @@
-export const prerender = true;
-
-/** @type {import('./$types').PageLoad} */
-export function load({ setHeaders }) {
-	setHeaders({
-		'cache-control': 'public, max-age=300'
-	});
-}
