--- packages/kit/test/apps/amp/src/routes/http-equiv/cache-control/+page.svelte
@@ -1,1 +1,0 @@
-<h1>the cache-control headers should be removed from this page</h1>
