--- packages/kit/test/apps/amp/src/routes/invalid/+page.svelte
@@ -1,2 +1,0 @@
-<!-- svelte-ignore a11y-invalid-attribute -->
-<a href="javascript:void(0);">invalid</a>
