--- packages/kit/test/apps/amp/src/routes/invalid/has-stylesheet/+page.svelte
@@ -1,3 +1,0 @@
-<svelte:head>
-	<link rel="stylesheet" href="https://example.com/whatever" />
-</svelte:head>
