--- packages/kit/test/apps/amp/src/routes/origin.json/+server.js
@@ -1,6 +1,0 @@
-import { json } from '@sveltejs/kit';
-
-/** @type {import('@sveltejs/kit').RequestHandler} */
-export function GET({ url }) {
-	return json({ origin: url.origin });
-}
