--- packages/kit/test/apps/amp/src/routes/styles/Unused.svelte
@@ -1,7 +1,0 @@
-<p class="shouty">this component will never be rendered</p>
-
-<style>
-	p {
-		color: #40b3ff;
-	}
-</style>
