--- packages/kit/test/apps/amp/src/routes/styles/imported.css
@@ -1,3 +1,0 @@
-.shouty {
-	text-transform: uppercase;
-}
