--- packages/kit/test/apps/amp/src/routes/valid.json/+server.js
@@ -1,5 +1,0 @@
-import { json } from '@sveltejs/kit';
-
-export function GET() {
-	return json({ answer: 42 });
-}
