--- packages/kit/test/apps/amp/src/routes/valid/+page.js
@@ -1,6 +1,0 @@
-/** @type {import('@sveltejs/kit').Load} */
-export async function load({ fetch }) {
-	const res = await fetch('/valid.json');
-	const { answer } = await res.json();
-	return { answer };
-}
