+++ packages/kit/test/apps/no-csr/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "test-amp",
+	"name": "test-no-csr",
 	"private": true,
 	"version": "0.0.1",
 	"scripts": {
@@ -8,15 +8,12 @@
 		"preview": "vite preview",
 		"prepare": "svelte-kit sync",
 		"check": "svelte-kit sync && tsc && svelte-check",
-		"test": "pnpm test:dev && pnpm test:build",
-		"test:dev": "DEV=true playwright test",
+		"test": "pnpm test:build",
 		"test:build": "playwright test"
 	},
 	"devDependencies": {
-		"@sveltejs/amp": "workspace:^",
 		"@sveltejs/kit": "workspace:^",
 		"@sveltejs/vite-plugin-svelte": "catalog:",
-		"dropcss": "catalog:",
 		"svelte": "catalog:",
 		"svelte-check": "catalog:",
 		"typescript": "catalog:",
