+++ packages/kit/test/apps/no-csr/playwright.config.js
@@ -1,0 +1,1 @@
+export { config as default } from '../../utils.js';
