+++ packages/kit/test/apps/no-csr/src/app.d.ts
