+++ packages/kit/test/apps/no-csr/src/app.html
@@ -1,9 +1,8 @@
 <!doctype html>
-<html lang="en" amp>
+<html lang="en">
 	<head>
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<link rel="canonical" href="https://example.com" />
 		%sveltekit.head%
 	</head>
 	<body>
