+++ packages/kit/test/apps/no-csr/src/routes/+layout.js
