+++ packages/kit/test/apps/no-csr/src/routes/assets/+page.svelte
