+++ packages/kit/test/apps/no-csr/src/routes/assets/favicon.png
