+++ packages/kit/test/apps/no-csr/static/asset.json
