+++ packages/kit/test/apps/no-csr/tsconfig.json
